refactor(json-crdt-peritext-ui): 💡 cleanup undo/redo controller code
